+++ apps/desktop/src/plugins/hermes-bots/bot-row.tsx
@@ -284,7 +284,7 @@ export function BotRow({ bot, onDelete, onEdit, onGroup, showHandle }: BotRowPro
     <ContextMenu>
       <ContextMenuTrigger asChild>{row}</ContextMenuTrigger>
       <ContextMenuContent>
-        <ContextMenuItem onSelect={() => void openRosterBot(bot, { canonical: true })}>
+        <ContextMenuItem onSelect={() => void openRosterBot(bot)}>
           {b.bot.openBotChat}
         </ContextMenuItem>
         <ContextMenuSeparator />
