+++ apps/desktop/src/plugins/hermes-bots/plugin.tsx
@@ -551,10 +551,10 @@ export default {
                 return
               }
 
-              // A claim without a registry id is a fronted non-canonical tab
-              // (focusExistingBotTab / the draft fallback): re-resolving the
-              // canonical chat here would open the Bot Chat the user has
-              // closed. Its tile recovers on the next send like any tab.
+              // A claim without a registry id is the legacy newChat draft
+              // fallback: re-resolving the canonical chat here would replace
+              // a draft the user is typing into. Its tile recovers on the next
+              // send like any tab.
               if (!claim.openedRegistryId) {
                 return
               }
